refactor: rename helpers to shared utils
