+++ functions/get-redeem-code.ts
@@ -2,7 +2,7 @@ import { verifyMessage } from "@ethersproject/wallet";
 import { getGiftCardOrderId, getMessageToSign } from "../shared/helpers";
 import { getRedeemCodeParamsSchema } from "../shared/api-types";
 import { getTransactionFromOrderId } from "./get-order";
-import { commonHeaders, getAccessToken, getReloadlyApiBaseUrl } from "./utils/helpers";
+import { commonHeaders, getAccessToken, getReloadlyApiBaseUrl } from "./utils/shared";
 import { AccessToken, Context, ReloadlyFailureResponse, ReloadlyRedeemCodeResponse } from "./utils/types";
 import { validateEnvVars, validateRequestMethod } from "./utils/validators";
 import { RedeemCode } from "../shared/types";
