+++ functions/utils/shared.ts
