+++ tests/fixtures/http-mocks.ts
@@ -13,7 +13,7 @@ import transaction0x33f4 from "./get-redeem-code/transaction-0x33f4.json";
 import card from "./get-redeem-code/card.json";
 import orderCard13959 from "./post-order/order-card-13959.json";
 import orderCard18597 from "./post-order/order-card-18597.json";
-import { RELOADLY_AUTH_URL, RELOADLY_PRODUCTION_API_URL, RELOADLY_SANDBOX_API_URL } from "../../functions/utils/helpers";
+import { RELOADLY_AUTH_URL, RELOADLY_PRODUCTION_API_URL, RELOADLY_SANDBOX_API_URL } from "../../functions/utils/shared";
 
 /**
  * Intercepts the routes and returns a custom payload
